# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Card Management System API
  description: >
    REST API for managing bank cards, users and transfers.
    This file is a static snapshot for reference. The authoritative,
    always-up-to-date spec is served live by the running application at:
      /v3/api-docs        (JSON)
      /swagger-ui.html     (interactive UI)
  version: 1.0.0
servers:
  - url: http://localhost:8080
tags:
  - name: Authentication
  - name: Users
  - name: Cards
  - name: Transfers
paths:
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user (first user becomes ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, email]
              properties:
                username: { type: string, minLength: 3, maxLength: 50 }
                password: { type: string, minLength: 6, maxLength: 100 }
                email: { type: string, format: email }
      responses:
        "201": { description: User registered, JWT returned }
        "400": { description: Username or email already taken }

  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate and obtain a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200": { description: JWT returned }
        "401": { description: Invalid credentials }

  /api/cards:
    get:
      tags: [Cards]
      summary: List current user's own cards (search + pagination)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [ACTIVE, BLOCKED, EXPIRED] }
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 20 }
      responses:
        "200": { description: Page of cards }
    post:
      tags: [Cards]
      summary: Create a new card for a user (admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cardNumber, ownerId, expirationDate]
              properties:
                cardNumber: { type: string, pattern: "^\\d{16}$" }
                ownerId: { type: integer }
                expirationDate: { type: string, format: date }
      responses:
        "201": { description: Card created }
        "403": { description: Forbidden - admin only }

  /api/cards/admin/all:
    get:
      tags: [Cards]
      summary: List all cards in the system (admin only)
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Page of all cards }

  /api/cards/{id}:
    get:
      tags: [Cards]
      summary: Get a card by id (owner or admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Card details }
        "403": { description: Not the owner }
        "404": { description: Card not found }
    delete:
      tags: [Cards]
      summary: Delete a card (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }

  /api/cards/{id}/block:
    patch:
      tags: [Cards]
      summary: Block a card (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Card blocked }

  /api/cards/{id}/activate:
    patch:
      tags: [Cards]
      summary: Activate a card (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Card activated }
        "400": { description: Card is expired, cannot activate }

  /api/cards/{id}/request-block:
    patch:
      tags: [Cards]
      summary: Request blocking of own card
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Block requested / applied }
        "403": { description: Not your card }

  /api/transfers:
    post:
      tags: [Transfers]
      summary: Transfer money between the current user's own cards
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fromCardId, toCardId, amount]
              properties:
                fromCardId: { type: integer }
                toCardId: { type: integer }
                amount: { type: number, format: decimal, minimum: 0.01 }
      responses:
        "200": { description: Transfer completed }
        "422": { description: Insufficient funds }
        "403": { description: Not your cards }

  /api/users:
    get:
      tags: [Users]
      summary: List all users (admin only)
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Page of users }

  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get a user by id (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: User details }
    delete:
      tags: [Users]
      summary: Delete a user (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT